refactor: (#58) Main auth modal 구조 분리
